create viewmodel module and extract some methods
