+++ rxlifecycleowner-viewmodel/gradle.properties
@@ -1,0 +1,3 @@
+POM_ARTIFACT_ID=rxlifecyclerowner-viewmodel
+POM_NAME=RxLifecycleOwner (ViewModel)
+POM_PACKAGING=aar
