+++ rxlifecycleowner/src/main/java/com/github/satoshun/io/reactivex/lifecycleowner/subscribers.kt
@@ -135,61 +135,3 @@ fun Completable.subscribeOf(
           .doOnTerminate { lifecycle.removeObserver(observer) }
           .subscribe(onComplete, onError)
     }
-
-@MainThread
-fun <T : Any> Flowable<T>.subscribeOf(
-    viewModel: RxViewModel,
-    onNext: (T) -> Unit = onNextEmpty,
-    onError: (Throwable) -> Unit = onErrorEmpty,
-    onComplete: () -> Unit = onCompleteEmpty
-): Disposable {
-  val disposable = subscribe(onNext, onError, onComplete)
-  viewModel.addDisposable(disposable)
-  return disposable
-}
-
-@MainThread
-fun <T : Any> Observable<T>.subscribeOf(
-    viewModel: RxViewModel,
-    onNext: (T) -> Unit = onNextEmpty,
-    onError: (Throwable) -> Unit = onErrorEmpty,
-    onComplete: () -> Unit = onCompleteEmpty
-): Disposable {
-  val disposable = subscribe(onNext, onError, onComplete)
-  viewModel.addDisposable(disposable)
-  return disposable
-}
-
-@MainThread
-fun <T : Any> Single<T>.subscribeOf(
-    viewModel: RxViewModel,
-    onSuccess: (T) -> Unit = onNextEmpty,
-    onError: (Throwable) -> Unit = onErrorEmpty
-): Disposable {
-  val disposable = subscribe(onSuccess, onError)
-  viewModel.addDisposable(disposable)
-  return disposable
-}
-
-@MainThread
-fun <T : Any> Maybe<T>.subscribeOf(
-    viewModel: RxViewModel,
-    onSuccess: (T) -> Unit = onNextEmpty,
-    onError: (Throwable) -> Unit = onErrorEmpty,
-    onComplete: () -> Unit = onCompleteEmpty
-): Disposable {
-  val disposable = subscribe(onSuccess, onError, onComplete)
-  viewModel.addDisposable(disposable)
-  return disposable
-}
-
-@MainThread
-fun Completable.subscribeOf(
-    viewModel: RxViewModel,
-    onComplete: () -> Unit = onCompleteEmpty,
-    onError: (Throwable) -> Unit = onErrorEmpty
-): Disposable {
-  val disposable = subscribe(onComplete, onError)
-  viewModel.addDisposable(disposable)
-  return disposable
-}
