+++ sample/src/main/java/com/github/satoshun/io/reactivex/lifecycleowner/sample/MainActivity.kt
@@ -4,8 +4,9 @@ import android.arch.lifecycle.ViewModelProviders
 import android.os.Bundle
 import android.support.v7.app.AppCompatActivity
 import android.util.Log
-import com.github.satoshun.io.reactivex.lifecycleowner.CompositeDisposableViewModel
 import com.github.satoshun.io.reactivex.lifecycleowner.subscribeOf
+import com.github.satoshun.io.reactivex.lifecycleowner.viewmodel.CompositeDisposableViewModel
+import com.github.satoshun.io.reactivex.lifecycleowner.viewmodel.subscribeOf
 import io.reactivex.Observable
 import io.reactivex.subjects.ReplaySubject
 import java.util.concurrent.TimeUnit
